сопоставление changeset и entity

+++ word-map/src/main/java/com/margot/word_map/dto/response/LetterResponse.java
@@ -24,8 +24,8 @@ public class LetterResponse {
     private LetterType type;
 
     @Schema(description = "Множитель бонуса за букву", example = "1")
-    private Short multiplier;
+    private Integer multiplier;
 
     @Schema(description = "Вес буквы (частота выпадения)", example = "6")
-    private Short weight;
+    private Integer weight;
 }
